backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
psycopg[binary]==3.2.3
pytest==8.3.3

